fix signed unsigned comparison

+++ src/upnp/quirks.cc
@@ -55,7 +55,7 @@ long Quirks::checkFlags(long flags) const
 
 bool Quirks::hasFlag(long flag) const
 {
-    return pClientInfo && (pClientInfo->flags & flag) == flag;
+    return pClientInfo && (static_cast<long>(pClientInfo->flags) & flag) == flag;
 }
 
 std::string Quirks::getGroup() const
